chore: move to `ic_canonical_state`
