+++ rs/canonical_state/tree_hash/src/lib.rs
@@ -1,3 +1,2 @@
 pub mod hash_tree;
 pub mod lazy_tree;
-pub mod witness;
